refactor(ast): single JS/TS entry package over facade and engine
Contributors kept picking between two confusingly named top-level
modules for JS/TS analysis: ast_js_ts (scanning facade with regex
fallback) and js_ts_ast (tree-sitter engine). Consolidate both under
one public package entry, agent_bom.ast.js_ts:

- move the facade to agent_bom/ast/js_ts/facade.py and the engine to
  agent_bom/ast/js_ts/engine.py; the package __init__ re-exports the
  scan entry points (scan_js_ts_file, build_js_ts_flow_findings,
  build_js_ts_dependency_symbol_reach) plus analyze_js_ts_block and
  the engine dataclasses, with public aliases for helpers that grew
  up module-private (JS_TS_EXTS, js_ts_function_key,
  npm_package_from_module)
- keep thin shims at the old module paths that delegate attribute
  lookups and emit a DeprecationWarning pointing at the package
- update ast_analyzer, parsers/skill_audit_behavior, and tests to
  import from the package
- the regex fallback when tree-sitter is unavailable is preserved and
  now covered by an explicit regression test

Verification: tests/test_ast_js_ts_package.py (7 new tests, red
before the package existed), 278-test AST/SAST/skill subset, full
pytest collection (15007 tests), ruff check src tests, mypy
src/agent_bom (762 files), agent-bom agents --demo --offline smoke,
and a wheel build asserting agent_bom/ast/js_ts ships in the
artifact.
