+++ src/agent_bom/ast/__init__.py
@@ -1,0 +1,7 @@
+"""Language AST analysis packages.
+
+Consolidation home for per-language AST analysis. Each language lives in
+its own subpackage with a single public entry point (start with
+``agent_bom.ast.js_ts``); the remaining top-level ``ast_*`` modules migrate
+here over time.
+"""
