+++ tests/test_js_ts_ast.py
@@ -4,7 +4,7 @@ pytest.importorskip("tree_sitter")
 pytest.importorskip("tree_sitter_javascript")
 pytest.importorskip("tree_sitter_typescript")
 
-from agent_bom.js_ts_ast import analyze_js_ts_block
+from agent_bom.ast.js_ts import analyze_js_ts_block
 
 
 def test_analyze_js_ts_block_resolves_named_import_alias():
